static generation of og.png images using a serverside route
